Safari-609.1.11.1/LayoutTests/http/tests/security/xssAuditor/img-onerror-accented-char-expected.txt

CONSOLE MESSAGE: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?test=/security/xssAuditor/img-onerror-accented-char.html&q=%3Cimg%20src=%C3%A4%20onerror=alert(%27%C3%A4%27)%3E' because its source code was found within the request. The auditor was enabled because the server did not send an 'X-XSS-Protection' header.
